+++ include/geode/inspector/criterion/internal/colocation_impl.h
@@ -30,7 +30,7 @@
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
         /*!
          * Implementation of the inspection of the colocation of a Mesh
@@ -50,5 +50,5 @@ namespace geode
         private:
             const Mesh& mesh_;
         };
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
