+++ src/geode/inspector/CMakeLists.txt
@@ -27,18 +27,18 @@ add_geode_library(
         "criterion/section_meshes_inspector.cpp"
         "criterion/adjacency/surface_adjacency.cpp"
         "criterion/adjacency/solid_adjacency.cpp"
-        "criterion/private/component_meshes_adjacency.cpp"
+        "criterion/internal/component_meshes_adjacency.cpp"
         "criterion/adjacency/section_meshes_adjacency.cpp"
         "criterion/adjacency/brep_meshes_adjacency.cpp"
-        "criterion/private/colocation_impl.cpp"
+        "criterion/internal/colocation_impl.cpp"
         "criterion/colocation/pointset_colocation.cpp"
         "criterion/colocation/edgedcurve_colocation.cpp"
         "criterion/colocation/surface_colocation.cpp"
         "criterion/colocation/solid_colocation.cpp"
         "criterion/colocation/unique_vertices_colocation.cpp"
         "criterion/colocation/component_meshes_colocation.cpp"
-        "criterion/private/degeneration_impl.cpp"
-        "criterion/private/component_meshes_degeneration.cpp"
+        "criterion/internal/degeneration_impl.cpp"
+        "criterion/internal/component_meshes_degeneration.cpp"
         "criterion/degeneration/edgedcurve_degeneration.cpp"
         "criterion/degeneration/surface_degeneration.cpp"
         "criterion/degeneration/solid_degeneration.cpp"
@@ -52,7 +52,7 @@ add_geode_library(
         "criterion/manifold/solid_vertex_manifold.cpp"
         "criterion/manifold/solid_edge_manifold.cpp"
         "criterion/manifold/solid_facet_manifold.cpp"
-        "criterion/private/component_meshes_manifold.cpp"
+        "criterion/internal/component_meshes_manifold.cpp"
         "criterion/manifold/section_meshes_manifold.cpp"
         "criterion/manifold/brep_meshes_manifold.cpp"
         "topology/brep_topology.cpp"
@@ -64,7 +64,7 @@ add_geode_library(
         "topology/section_corners_topology.cpp"
         "topology/section_lines_topology.cpp"
         "topology/section_surfaces_topology.cpp"
-        "topology/private/topology_helpers.cpp"
+        "topology/internal/topology_helpers.cpp"
         "section_inspector.cpp"
         "brep_inspector.cpp"
         "pointset_inspector.cpp"
@@ -110,12 +110,12 @@ add_geode_library(
         "edgedcurve_inspector.h"
         "surface_inspector.h"
         "solid_inspector.h"
-    PRIVATE_HEADERS
-        "criterion/private/colocation_impl.h"
-        "criterion/private/component_meshes_adjacency.h"
-        "criterion/private/component_meshes_degeneration.h"
-        "criterion/private/component_meshes_manifold.h"
-        "criterion/private/degeneration_impl.h"
+    INTERNAL_HEADERS
+        "criterion/internal/colocation_impl.h"
+        "criterion/internal/component_meshes_adjacency.h"
+        "criterion/internal/component_meshes_degeneration.h"
+        "criterion/internal/component_meshes_manifold.h"
+        "criterion/internal/degeneration_impl.h"
         "topology/brep_corners_topology.h"
         "topology/brep_lines_topology.h"
         "topology/brep_surfaces_topology.h"
@@ -123,7 +123,7 @@ add_geode_library(
         "topology/section_corners_topology.h"
         "topology/section_lines_topology.h"
         "topology/section_surfaces_topology.h"
-        "topology/private/topology_helpers.h"
+        "topology/internal/topology_helpers.h"
     PUBLIC_DEPENDENCIES
         OpenGeode::basic
     PRIVATE_DEPENDENCIES
