+++ src/geode/inspector/criterion/adjacency/brep_meshes_adjacency.cpp
@@ -30,7 +30,7 @@
 #include <geode/model/representation/core/brep.h>
 
 #include <geode/inspector/criterion/adjacency/solid_adjacency.h>
-#include <geode/inspector/criterion/private/component_meshes_adjacency.h>
+#include <geode/inspector/criterion/internal/component_meshes_adjacency.h>
 
 namespace geode
 {
