+++ src/geode/inspector/criterion/degeneration/section_meshes_degeneration.cpp
@@ -30,7 +30,7 @@
 #include <geode/model/representation/core/section.h>
 
 #include <geode/inspector/criterion/degeneration/solid_degeneration.h>
-#include <geode/inspector/criterion/private/component_meshes_degeneration.h>
+#include <geode/inspector/criterion/internal/component_meshes_degeneration.h>
 
 namespace geode
 {
