+++ src/geode/inspector/topology/brep_lines_topology.cpp
@@ -36,7 +36,7 @@
 #include <geode/model/mixin/core/surface.h>
 #include <geode/model/representation/core/brep.h>
 
-#include <geode/inspector/topology/private/topology_helpers.h>
+#include <geode/inspector/topology/internal/topology_helpers.h>
 
 namespace geode
 {
@@ -172,7 +172,7 @@ namespace geode
                         embedding.id().string() + "'." );
                 }
                 if( embedding.type() == Block3D::component_type_static()
-                    && !detail::brep_blocks_are_meshed( brep_ ) )
+                    && !internal::brep_blocks_are_meshed( brep_ ) )
                 {
                     continue;
                 }
@@ -199,16 +199,16 @@ namespace geode
         BRepLinesTopology::vertex_is_part_of_invalid_single_line(
             index_t unique_vertex_index ) const
     {
-        const auto line_uuids = detail::components_uuids(
+        const auto line_uuids = internal::components_uuids(
             brep_, unique_vertex_index, Line3D::component_type_static() );
         if( line_uuids.size() != 1 )
         {
             return std::nullopt;
         }
         const auto& line_id = line_uuids[0];
-        const auto surface_uuids = detail::components_uuids(
+        const auto surface_uuids = internal::components_uuids(
             brep_, unique_vertex_index, Surface3D::component_type_static() );
-        const auto block_uuids = detail::components_uuids(
+        const auto block_uuids = internal::components_uuids(
             brep_, unique_vertex_index, Block3D::component_type_static() );
         if( surface_uuids.size() == 1 )
         {
@@ -229,7 +229,7 @@ namespace geode
         }
         else if( surface_uuids.empty() )
         {
-            if( !detail::brep_blocks_are_meshed( brep_ ) )
+            if( !internal::brep_blocks_are_meshed( brep_ ) )
             {
                 return std::nullopt;
             }
@@ -313,7 +313,7 @@ namespace geode
                                    " is a line without mesh." ) );
             }
 
-            auto line_result = detail::
+            auto line_result = internal::
                 brep_component_vertices_not_associated_to_unique_vertices(
                     brep_, line.component_id(), line.mesh() );
             if( line_result.nb_issues() != 0 )
